formatting + comment, eliminate redundant fnc

+++ src/contracts/middleware/DataLayr/DataLayrServiceManagerStorage.sol
@@ -183,7 +183,7 @@ abstract contract DataLayrServiceManagerStorage is IDataLayrServiceManager, ISer
      * @notice mapping between the total service fee that would be paid out in the 
      *         corresponding assertion of data into DataLayr 
      */
-    mapping(uint64 => uint256) public dumpNumberToFee;
+    mapping(uint32 => uint256) public dumpNumberToFee;
 
 
     /**
